fix(Modal): set display block always, fixes #3399 (#3571)
We started setting `display: block` to div.modal at onEnter timing of Transition at Pull #3187. But when we set `animation={false}` to Modal, the Transition object is not created and therefore onEnter (=handleEnter) is not called. This causes empty modal when animation set `false` (I think this is the issue #3399 ). This PR tries to fix this behavior by setting the modal's display style to `block` always.

I think this doesn't break existing behavior because as @jquense says
> we don't allow modals to be in the DOM prior to showing 

the modal's dom element doesn't exist before rendering. So I think it's safe to set display always `block`.