chore(AccordionCollapse): add proper display name
The display name is also being used as a selector in the
Accordion spec re-write, so it's important that this display name
matches up with the name specified in the spec.

Unfortunately, a lot of our components are relying on a display
name for testing, which seems to be testing implementation
details. Hopefully we can migrate away from using display names
as selectors in our unit tests.