feat(netlify): more resilient netlify docs alert implementation
Uses the already built-in netlify configuration variables that
will be injected by the netlify build environment, which should
allow use to not have to configure anything extra to get this
feature to work properly.

Also adds some more information about the deployment, specifying
whether or not it's a pull request (and redirects to the right
pr link), or if it's being deployed from a specific branch.